+++ client/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Dashboard from './components/materials/testDashboard/Dashboard';
+import Dashboard from './components/materials/adminDashboard/Dashboard';
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Test from './components/test';
 const App = () => (
